move bootnodes to primitives, make bootnodes method on chain
