+++ crates/net/network/src/lib.rs
@@ -55,9 +55,10 @@
 //!
 //! ```
 //! # async fn launch() {
-//! use reth_network::config::{rng_secret_key, mainnet_nodes};
+//! use reth_network::config::rng_secret_key;
 //! use reth_network::{NetworkConfig, NetworkManager};
 //! use reth_provider::test_utils::NoopProvider;
+//! use reth_primitives::mainnet_nodes;
 //!
 //! // This block provider implementation is used for testing purposes.
 //! let client = NoopProvider::default();
@@ -84,7 +85,7 @@
 //! ```
 //! use reth_provider::test_utils::NoopProvider;
 //! use reth_transaction_pool::TransactionPool;
-//! use reth_discv4::bootnodes::mainnet_nodes;
+//! use reth_primitives::mainnet_nodes;
 //! use reth_network::config::rng_secret_key;
 //! use reth_network::{NetworkConfig, NetworkManager};
 //! async fn launch<Pool: TransactionPool>(pool: Pool) {
